+++ src/Exceptions/InvalidValueObjectsArgumentException.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * This file is part of ZeroBoiler, licensed under the MIT license.
  */
@@ -14,8 +13,10 @@ namespace ZeroBoiler\ValueObjects\Exceptions;
  * Replaces generic \InvalidArgumentException with a domain-specific
  * exception for better error handling and debugging.
  *
- * @see ValueObjectsException
+ * @see \\ZeroBoiler\\ValueObjects\\Exceptions\\ValueObjectsException
  *
+ * @see \ZeroBoiler\ValueObjects\Exceptions\ValueObjectsException
+ * @see \ZeroBoiler\ValueObjects\Exceptions\ValueObjectsRuntimeException
  * @since 1.26.0
  */
 final class InvalidValueObjectsArgumentException extends ValueObjectsException